--- PR_SUMMARY.md
@@ -1,216 +1,0 @@
-# Pull Request: Add BacktestRunner for Automated Portfolio Simulation
-
-**Branch:** `claude/design-backtest-runner-011CUcC4gymNbdN3ckTuDccw`
-**Base:** `development`
-**Status:** ✅ Ready for Review
-
----
-
-## 🎯 Summary
-
-This PR introduces **BacktestRunner**, a high-level automation layer that makes Pyfolium significantly more ergonomic for end users by eliminating tedious manual period loops.
-
-## ✨ New Features
-
-### 1. **BacktestRunner** (`pyfolium/simulation.py`)
-Automates the complete backtest simulation loop with advanced features:
-
-```python
-# Simple usage - just 3 lines!
-runner = BacktestRunner(portfolio, strategy)
-result = runner.run()
-print(f"Final cash: ${result.portfolio.cash:,.2f}")
-```
-
-**Key Features:**
-- ✅ **Progress Reporting**: Optional `tqdm` progress bars (`progress=True`)
-- ✅ **Custom Hooks**: 5 event types for extensibility
-  - `period_start`, `period_end`, `backtest_start`, `backtest_end`, `error`
-- ✅ **Step-by-Step Control**: `runner.run_period()` for manual debugging
-- ✅ **Custom Period Ranges**: Run partial backtests
-- ✅ **Context Manager**: `with BacktestRunner(...) as runner:`
-- ✅ **Error Handling**: Catch, store, continue with warnings
-
-### 2. **Portfolio.clone()** (`pyfolium/core.py`)
-Deep copy portfolios for strategy comparison and optimization:
-
-```python
-portfolio1 = base_portfolio.clone()
-portfolio2 = base_portfolio.clone()
-
-result1 = BacktestRunner(portfolio1, strategy1).run()
-result2 = BacktestRunner(portfolio2, strategy2).run()
-```
-
-**Perfect for:**
-- Parameter optimization & sensitivity analysis
-- A/B testing strategies
-- Future: Parallel execution (Phase 2)
-
-### 3. **BacktestResult**
-Lean data container with execution metadata:
-- Portfolio & strategy references
-- Execution time, period counts
-- Error tracking
-- `success` property
-
-## 📦 What's Included
-
-### Code
-- `pyfolium/simulation.py` - BacktestRunner & BacktestResult (400+ LOC)
-- `pyfolium/core.py` - Portfolio.clone() method
-- `pyfolium/__init__.py` - Export new classes
-- Merged latest development (Pydantic validation, data loaders)
-
-### Tests
-- ✅ 20/20 tests passing
-- `tests/simulation/test_backtest_runner_minimal.py` - Core functionality
-- `tests/core/test_portfolio_clone_minimal.py` - Clone tests
-- Comprehensive test coverage for edge cases
-
-### Documentation
-- Updated README with comprehensive examples
-- `examples/backtest_runner_example.py` - 7 complete usage examples:
-  1. Simple usage
-  2. Progress bars
-  3. Custom hooks for logging
-  4. Step-by-step execution
-  5. Comparing multiple strategies
-  6. Custom period ranges
-  7. Context manager pattern
-
-### Phase 2 Roadmap (Documented)
-- **ParameterSweep**: Grid search over strategy parameters
-- **BatchRunner**: Parallel execution of multiple backtests
-- **Optimization Integration**: scipy.optimize, optuna, etc.
-- **Result Caching**: Avoid re-running identical configurations
-- **Performance Metrics**: Built-in Sharpe ratio, drawdown calculations
-
-## 🔄 Changes from Development Base
-
-This PR is rebased on latest `development` branch and includes:
-- ✅ Pydantic validation for TaxConfig/FeeConfig
-- ✅ New `pyfolium/data.py` module with data loaders
-- ✅ Python 3.12 compatibility fixes
-- ✅ All new BacktestRunner functionality
-
-## 🧪 Testing
-
-All tests pass:
-```bash
-$ uv run pytest tests/simulation/test_backtest_runner_minimal.py tests/core/test_portfolio_clone_minimal.py
-================================ 20 passed in 0.81s ================================
-```
-
-**Test Coverage:**
-- BacktestRunner initialization & configuration ✓
-- Hooks registration & execution ✓
-- Error handling & recovery ✓
-- Portfolio cloning & independence ✓
-- Period range validation ✓
-
-## 📖 Usage Example
-
-### Before (tedious manual loop):
-```python
-for period in universe.get_period_index_range():
-    portfolio.collect_income()
-    strategy.step()
-    portfolio.update_history()
-    portfolio.advance_period()
-```
-
-### After (clean & powerful):
-```python
-runner = BacktestRunner(portfolio, strategy)
-result = runner.run(progress=True)
-print(f"Completed {result.total_periods} periods in {result.execution_time:.2f}s")
-```
-
-### Advanced Usage:
-```python
-# Custom hooks for logging
-def log_value(runner):
-    holdings = runner.portfolio.holdings.loc[runner.current_period]
-    prices = runner.asset_universe.price_matrix.loc[runner.current_period]
-    value = runner.portfolio.cash + (holdings * prices).sum()
-    print(f"Portfolio value: ${value:,.2f}")
-
-runner = BacktestRunner(portfolio, strategy)
-runner.register_hook('period_end', log_value)
-result = runner.run()
-```
-
-## 🎨 Design Decisions
-
-1. **Pattern Choice**: Object-oriented runner with hooks (Pattern 2)
-   - Most flexible for advanced users
-   - Simple for basic use cases
-   - Event-driven extensibility
-
-2. **Lean Results**: No built-in metrics calculations
-   - Users calculate from `portfolio.history`
-   - Keeps API surface small
-   - Maximum flexibility
-
-3. **Error Handling**: Store & continue with warnings
-   - Graceful degradation
-   - Full error log in `result.errors`
-   - Allows partial backtest completion
-
-4. **Clone for Optimization**: Simple deep copy method
-   - Enables future parallelization
-   - Independent portfolio instances
-   - Shared immutable AssetUniverse
-
-## 🚀 Breaking Changes
-
-**None!** This is purely additive. All existing code continues to work.
-
-## 📋 Checklist
-
-- [x] Code implemented and tested
-- [x] All tests passing (20/20)
-- [x] Documentation updated (README + examples)
-- [x] Code formatted with ruff
-- [x] Merged latest development branch
-- [x] Phase 2 roadmap documented
-
-## 🎯 Goals Achieved
-
-✅ Eliminate tedious manual period loops
-✅ Maintain Portfolio state machine integrity
-✅ Enable strategy comparison & optimization
-✅ Provide hooks for custom logic
-✅ Keep API simple for basic use cases
-✅ Make Pyfolium ergonomic for end users
-
----
-
-## 📝 Files Changed
-
-```
-examples/backtest_runner_example.py    | 360 ++++++++++++++++++++++++++++++
-pyfolium/__init__.py                   |  10 +-
-pyfolium/core.py                       |  50 +++++
-pyfolium/simulation.py                 | 405 +++++++++++++++++++++++++++++++++
-tests/core/test_portfolio_clone.py     | 295 ++++++++++++++++++++++++
-tests/core/test_portfolio_clone_minimal.py | 121 ++++++++++
-tests/simulation/__init__.py           |   1 +
-tests/simulation/test_backtest_runner.py | 445 +++++++++++++++++++++++++++++++++++
-tests/simulation/test_backtest_runner_minimal.py | 166 ++++++++++++++
-
-9 files changed, 1850 insertions(+), 3 deletions(-)
-```
-
-## 🔗 Links
-
-- **GitHub PR**: https://github.com/whnr/pyfolium/pull/new/claude/design-backtest-runner-011CUcC4gymNbdN3ckTuDccw
-- **Branch**: `claude/design-backtest-runner-011CUcC4gymNbdN3ckTuDccw`
-- **Compare**: `development...claude/design-backtest-runner-011CUcC4gymNbdN3ckTuDccw`
-
----
-
-**Ready for review!** 🚀
-
-The goal was to make Pyfolium ergonomic for end users while maintaining the robustness of the current state machine. Mission accomplished! ✨
